pyproject.toml
[project]
name = "traveldocumentsagent"
version = "0.1.0"
description = "Travel Documents Agent - Creates visual travel documents with local events and attractions using Google ADK"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.1.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "weasyprint>=62.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["travel_documents_agent*"]
